requirements.txt
discord.py==2.1.0
pytz==2022.6
sentry-sdk==1.12.0
pyyaml==6.0
prometheus_client==0.15.0
pillow==9.3.0
requests==2.28.1
aiohttp==3.8.3
setuptools==65.6.3
tortoise-orm==0.19.2
asyncmy==0.2.5
image-similarity-measures==0.3.5
opencv-python==4.6.0.66
numpy==1.23.5
aerich==0.7.1